requirements.txt
fastapi
uvicorn
pydantic
langchain-core
langchain-ollama
langchain
langchain-community
langchain-chroma
sentence-transformers
python-dotenv
pdfplumber
unstructured